requirements.txt
langgraph
langsmith
langchain[google-genai]
python-dotenv
ipython

